+++ src/main/java/Twofort/Commands/CommandList.java
@@ -1,4 +1,4 @@
-package Duke.Commands;
+package Twofort.Commands;
 
 public enum CommandList {
     BYE, LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, FIND, TAG, UNTAG
