+++ src/main/java/Twofort/Commands/DeadlineCommand.java
@@ -1,7 +1,7 @@
-package Duke.Commands;
+package Twofort.Commands;
 
-import Duke.Saver;
-import Duke.Tasks.TaskList;
+import Twofort.Saver;
+import Twofort.Tasks.TaskList;
 
 public class DeadlineCommand extends Command {
     private String name;
