+++ src/main/java/Twofort/Commands/UntagCommand.java
@@ -1,22 +1,22 @@
-package Duke.Commands;
+package Twofort.Commands;
 
-import Duke.Saver;
-import Duke.Tasks.TaskList;
+import Twofort.Saver;
+import Twofort.Tasks.TaskList;
 
 import java.util.ArrayList;
 
-public class TagCommand extends Command {
+public class UntagCommand extends Command {
     private int index;
     private ArrayList<String> tags;
 
-    public TagCommand(int index, ArrayList<String> tags) {
+    public UntagCommand(int index, ArrayList<String> tags) {
         this.index = index;
         this.tags = tags;
     }
 
     @Override
     public String run(TaskList taskList, Saver saver) {
-        String message = taskList.tagTask(index, tags);
+        String message = taskList.untagTask(index, tags);
         saver.save(taskList);
         return message;
