+++ src/main/java/Twofort/Launcher.java
@@ -1,4 +1,4 @@
-package Duke;
+package Twofort;
 
 import javafx.application.Application;
