+++ src/main/java/Twofort/Tasks/Task.java
@@ -1,4 +1,4 @@
-package Duke.Tasks;
+package Twofort.Tasks;
 
 import java.util.ArrayList;
