+++ src/main/java/Twofort/Tasks/ToDo.java
@@ -1,6 +1,4 @@
-package Duke.Tasks;
-
-import Duke.Tasks.Task;
+package Twofort.Tasks;
 
 /**
  * Class representing a ToDo task.
