+++ src/main/java/Twofort/graphics/DialogBox.java
@@ -1,4 +1,4 @@
-package Duke.graphics;
+package Twofort.graphics;
 
 import java.io.IOException;
 import java.util.Collections;
