+++ src/main/java/Twofort/graphics/MainWindow.java
@@ -1,6 +1,6 @@
-package Duke.graphics;
+package Twofort.graphics;
 
-import Duke.Duke;
+import Twofort.Twofort;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
@@ -21,19 +21,19 @@ public class MainWindow extends AnchorPane {
     @FXML
     private Button sendButton;
 
-    private Duke duke;
+    private Twofort twofort;
 
-    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
-    private Image dukeImage = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
-    private Image dukeErrorImage = new Image(this.getClass().getResourceAsStream("/images/DaDukeError.png"));
+    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/Scout.png"));
+    private Image dukeImage = new Image(this.getClass().getResourceAsStream("/images/Engineer.png"));
+    private Image dukeErrorImage = new Image(this.getClass().getResourceAsStream("/images/EngineerError.png"));
 
     @FXML
     public void initialize() {
         scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
     }
 
-    public void setDuke(Duke d) {
-        duke = d;
+    public void setTwofort(Twofort t) {
+        twofort = t;
         getWelcomeMessage();
     }
 
@@ -45,7 +45,7 @@ public class MainWindow extends AnchorPane {
     @FXML
     private void handleUserInput() {
         String input = userInput.getText();
-        String response = duke.getResponse(input);
+        String response = twofort.getResponse(input);
         if(response.split(" ",2)[0].equals("[E]")){ //error message
             dialogContainer.getChildren().addAll(
                     DialogBox.getUserDialog(input, userImage),
