+++ src/main/resources/images/Scout.png
