+++ src/main/resources/view/DialogBox.fxml
@@ -6,7 +6,7 @@
 
 <fx:root alignment="TOP_RIGHT" maxHeight="Infinity" prefWidth="400.0" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/" xmlns:fx="http://javafx.com/fxml/1">
     <children>
-        <Label fx:id="dialog" alignment="TOP_LEFT" contentDisplay="TOP" style="-fx-background-color: #222222; -fx-border-radius: 5px; -fx-background-radius: 5px; -fx-font-size: 14;-fx-padding: 10;" wrapText="true" text="Label" textFill="WHITE">
+        <Label fx:id="dialog" minHeight="-Infinity" alignment="TOP_LEFT" contentDisplay="TOP" style="-fx-background-color: #222222; -fx-border-radius: 5px; -fx-background-radius: 5px; -fx-font-size: 14;-fx-padding: 10;" wrapText="true" text="Label" textFill="WHITE">
         </Label>
         <ImageView fx:id="displayPicture" fitHeight="99.0" fitWidth="99.0" pickOnBounds="true" preserveRatio="true">
         </ImageView>
